Deferred reading states in SearchBarWithResults.kt made SearchBarWithResults take the component responsible for displaying the results as a parameter to make the hierarchy a bit easier to read and broken the Transparent search bar into smaller components
