--- app/src/main/java/com/trianglz/weatherapp/presentation/models/country/CountryUiModel.kt
@@ -1,9 +1,0 @@
-package com.trianglz.weatherapp.presentation.models.country
-
-import com.trianglz.weatherapp.presentation.models.resultitem.ResultItem
-
-data class CountryUiModel(
-    override val name: String,
-    val code: String,
-    val flag: String
-) : ResultItem
