+++ app/src/main/java/com/trianglz/weatherapp/presentation/models/result/ResultItem.kt
@@ -1,0 +1,5 @@
+package com.trianglz.weatherapp.presentation.models.result
+
+interface ResultItem {
+    val name: String
+}
