+++ app/src/main/java/com/trianglz/weatherapp/presentation/models/result/ResultUiModel.kt
@@ -1,0 +1,7 @@
+package com.trianglz.weatherapp.presentation.models.result
+
+data class ResultUiModel(
+    override val name: String,
+    val code: String,
+    val flag: String
+) : ResultItem
