--- app/src/main/java/com/trianglz/weatherapp/presentation/models/resultitem/ResultItem.kt
@@ -1,5 +1,0 @@
-package com.trianglz.weatherapp.presentation.models.resultitem
-
-interface ResultItem {
-    val name: String
-}
