+++ app/src/main/java/com/trianglz/weatherapp/presentation/models/searchbar/SearchBarState.kt
@@ -1,7 +1,8 @@
 package com.trianglz.weatherapp.presentation.models.searchbar
 
+import com.trianglz.weatherapp.presentation.models.result.ResultState
+
 data class SearchBarState(
     val placeHolder: String = "",
-    val status: SearchBarStatus = SearchBarStatus.Idle,
-    val noResultMessage: String = "",
+    val status: ResultState<Nothing> = ResultState.Idle(),
 )
