--- app/src/main/java/com/trianglz/weatherapp/presentation/models/searchbar/SearchBarStatus.kt
@@ -1,7 +1,0 @@
-package com.trianglz.weatherapp.presentation.models.searchbar
-
-sealed class SearchBarStatus {
-    object Loading : SearchBarStatus()
-    object Error : SearchBarStatus()
-    object Idle : SearchBarStatus()
-}
