+++ app/src/main/java/com/trianglz/weatherapp/presentation/ui/screentwo/ScreenTwoViewModel.kt
@@ -43,4 +43,3 @@ class ScreenTwoViewModel
         )
     }
 }
-
